pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "A CLI tool to compare configuration files"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["PyYAML"]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
